Move text utility files to ui-core module
