+++ feature-compose/src/commonMain/kotlin/com/tunjid/heron/compose/ComposeScreen.kt
@@ -50,7 +50,7 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import com.tunjid.heron.compose.ui.MediaUploadItems
-import com.tunjid.heron.timeline.utilities.links
+import com.tunjid.heron.ui.text.links
 import com.tunjid.heron.data.core.models.Post
 import com.tunjid.heron.data.core.models.Profile
 import com.tunjid.heron.data.core.models.contentDescription
